Rework the whoami and login tests to use wasmer.wtf
