+++ tests/integration/cli/tests/login.rs
@@ -1,53 +1,61 @@
-use anyhow::bail;
+use assert_cmd::prelude::OutputAssertExt;
+use predicates::str::contains;
+use tempfile::TempDir;
 
 use std::process::Command;
 use wasmer_integration_tests_cli::get_wasmer_path;
 
 #[test]
-fn login_works() -> anyhow::Result<()> {
+fn login_works() {
+    let wasmer_dir = TempDir::new().unwrap();
+
     // running test locally: should always pass since
     // developers don't have access to WAPM_DEV_TOKEN
     if std::env::var("GITHUB_TOKEN").is_err() {
-        return Ok(());
+        return;
     }
     let wapm_dev_token = std::env::var("WAPM_DEV_TOKEN").expect("WAPM_DEV_TOKEN env var not set");
     // Special case: GitHub secrets aren't visible to outside collaborators
     if wapm_dev_token.is_empty() {
-        return Ok(());
+        return;
     }
-    // FIXME: Change the registry to wasmer.wtf and all the associated terms with WAPM
-    let output = Command::new(get_wasmer_path())
+    let assert = Command::new(get_wasmer_path())
         .arg("login")
-        .arg("--registry")
-        .arg("wapm.dev")
+        .arg("--registry=wasmer.wtf")
         .arg(wapm_dev_token)
-        .output()?;
+        .env("WASMER_DIR", wasmer_dir.path())
+        .assert();
 
-    let stdout = std::str::from_utf8(&output.stdout)
-        .expect("stdout is not utf8! need to handle arbitrary bytes");
+    assert
+        .success()
+        .stdout(contains(r#"Login for Wasmer user "ciuser" saved"#));
+}
 
-    let stderr = std::str::from_utf8(&output.stderr)
-        .expect("stderr is not utf8! need to handle arbitrary bytes");
+#[test]
+fn run_whoami_works() {
+    let wasmer_dir = TempDir::new().unwrap();
 
-    if !output.status.success() {
-        bail!(
-            "wasmer login failed with: stdout: {}\n\nstderr: {}",
-            stdout,
-            stderr
-        );
+    // running test locally: should always pass since
+    // developers don't have access to WAPM_DEV_TOKEN
+    if std::env::var("GITHUB_TOKEN").is_err() {
+        return;
     }
 
-    let stdout_output = std::str::from_utf8(&output.stdout).unwrap();
-    let expected = "Done!\n✅ Login for Wasmer user \"ciuser\" saved\n";
-    if stdout_output != expected {
-        println!("expected:");
-        println!("{expected}");
-        println!("got:");
-        println!("{stdout}");
-        println!("-----");
-        println!("{stderr}");
-        panic!("stdout incorrect");
+    let ciuser_token = std::env::var("WAPM_DEV_TOKEN").expect("no CIUSER / WAPM_DEV_TOKEN token");
+    // Special case: GitHub secrets aren't visible to outside collaborators
+    if ciuser_token.is_empty() {
+        return;
     }
 
-    Ok(())
+    let assert = Command::new(get_wasmer_path())
+        .arg("whoami")
+        .arg("--registry=wasmer.wtf")
+        .env("WASMER_TOKEN", &ciuser_token)
+        .env("WASMER_DIR", wasmer_dir.path())
+        .assert()
+        .success();
+
+    assert.stdout(
+        "logged into registry \"https://registry.wasmer.wtf/graphql\" as user \"ciuser\"\n",
+    );
 }
